convert testArgs to ts and typescript tests to ava also fixed some typing problems that became apparent because of it
