+++ tsconfig.json
@@ -18,6 +18,6 @@
     "declarationMap": true
   },
   "files": ["./ts/index.ts"],
-  "include": ["./ts/tests.ts"],
+  "include": ["./test/typescript.test.ts"],
   "exclude": ["./lib/*.js"]
 }
